homePage, registration, login added on frontend. Validation added on backend

+++ backend/Cineverse/Cineverse.Application/DTOs/Reviews/CreateReviewRequest.cs
@@ -8,7 +8,7 @@ namespace Cineverse.Application.DTOs.Reviews
 {
     public record CreateReviewRequest(
         int TmdbMovieId,
-        int Rating,
+        double Rating,
         string Content
     );    
 }
